+++ plugin-build/src/agp74/kotlin/io/sentry/android/gradle/AGP74Compat.kt
@@ -19,6 +19,7 @@ import org.gradle.api.file.Directory
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.FileCollection
 import org.gradle.api.provider.Provider
+import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.TaskProvider
 
 data class AndroidVariant74(
@@ -77,6 +78,7 @@ fun <T : InstrumentationParameters> configureInstrumentationFor74(
     classVisitorFactoryImplClass: Class<out AsmClassVisitorFactory<T>>,
     scope: InstrumentationScope,
     mode: FramesComputationMode,
+    excludes: SetProperty<String>,
     instrumentationParamsConfig: (T) -> Unit
 ) {
     variant.instrumentation.transformClassesWith(
@@ -85,6 +87,7 @@ fun <T : InstrumentationParameters> configureInstrumentationFor74(
         instrumentationParamsConfig
     )
     variant.instrumentation.setAsmFramesComputationMode(mode)
+    variant.instrumentation.excludes.set(excludes)
 }
 
 fun onVariants74(